[Tech] delete reportingState and use only MultiReportings slice
