+++ frontend/src/domain/shared_slices/index.js
@@ -14,7 +14,6 @@ import { multiReportingsSliceReducer } from './MultiReportings'
 import { regulatorySlicePersistedReducer } from './Regulatory'
 import { regulatoryMetadataSliceReducer } from './RegulatoryMetadata'
 import { reportingFiltersPersistedReducer } from './ReportingsFilters'
-import { reportingStateSliceReducer } from './ReportingState'
 import { selectedAmpSlicePersistedReducer } from './SelectedAmp'
 import { semaphoresPersistedReducer } from './SemaphoresSlice'
 import { ampsAPI, ampsErrorLoggerMiddleware } from '../../api/ampsAPI'
@@ -51,7 +50,6 @@ export const homeReducers = combineReducers({
   [infractionsAPI.reducerPath]: infractionsAPI.reducer,
   [semaphoresAPI.reducerPath]: semaphoresAPI.reducer,
   reportingFilters: reportingFiltersPersistedReducer,
-  reportingState: reportingStateSliceReducer,
   [reportingsAPI.reducerPath]: reportingsAPI.reducer,
   selectedAmp: selectedAmpSlicePersistedReducer,
   semaphoresSlice: semaphoresPersistedReducer,
