+++ frontend/src/domain/use_cases/reportings/addReporting.ts
@@ -1,8 +1,6 @@
-import { saveReportingInMultiReportingsState } from './saveReportingInMultiReportingsState'
 import { getReportingInitialValues, createIdForNewReporting } from '../../../features/Reportings/utils'
-import { setReportingFormVisibility } from '../../shared_slices/Global'
+import { setReportingFormVisibility, ReportingContext, VisibilityState } from '../../shared_slices/Global'
 import { multiReportingsActions } from '../../shared_slices/MultiReportings'
-import { ReportingContext, VisibilityState } from '../../shared_slices/ReportingState'
 
 import type { Reporting } from '../../entities/reporting'
 
@@ -10,29 +8,20 @@ export const addReporting =
   (reportingContext: ReportingContext, partialReporting?: Partial<Reporting> | undefined) =>
   async (dispatch, getState) => {
     const {
-      multiReportings: { selectedReportings },
-      reportingState: { reportingState }
+      multiReportings: { selectedReportings }
     } = getState()
-    let reportings = [...selectedReportings]
+    const reportings = { ...selectedReportings }
 
-    // first we want to save the active reporting in multiReportings state
-    if (reportingState) {
-      reportings = await dispatch(saveReportingInMultiReportingsState())
-    }
     const id = createIdForNewReporting(reportings)
 
-    const updatedReportings = [
-      ...reportings,
-      {
-        context: reportingContext,
-        isFormDirty: false,
-        reporting: getReportingInitialValues({ createdAt: new Date().toISOString(), id, ...partialReporting })
-      }
-    ]
+    const newReporting = {
+      context: reportingContext,
+      isFormDirty: false,
+      reporting: getReportingInitialValues({ createdAt: new Date().toISOString(), id, ...partialReporting })
+    }
 
-    await dispatch(
-      multiReportingsActions.setSelectedReportings({ activeReportingId: id, selectedReportings: updatedReportings })
-    )
+    await dispatch(multiReportingsActions.setReporting(newReporting))
+    // await dispatch(multiReportingsActions.setActiveReportingId(id))
 
     await dispatch(
       setReportingFormVisibility({
