+++ frontend/src/domain/use_cases/reportings/duplicateReporting.ts
@@ -1,14 +1,13 @@
 import { reportingsAPI } from '../../../api/reportingsAPI'
 import { getReportingInitialValues, createIdForNewReporting } from '../../../features/Reportings/utils'
-import { setReportingFormVisibility, setToast } from '../../shared_slices/Global'
+import { setReportingFormVisibility, setToast, ReportingContext, VisibilityState } from '../../shared_slices/Global'
 import { multiReportingsActions } from '../../shared_slices/MultiReportings'
-import { ReportingContext, VisibilityState } from '../../shared_slices/ReportingState'
 
 export const duplicateReporting = (reportingId: number) => async (dispatch, getState) => {
   const {
     multiReportings: { selectedReportings }
   } = getState()
-  const reportings = [...selectedReportings]
+  const reportings = { ...selectedReportings }
 
   const reportingToDuplicate = reportingsAPI.endpoints.getReporting
   try {
@@ -16,18 +15,14 @@ export const duplicateReporting = (reportingId: number) => async (dispatch, getS
     if ('data' in response) {
       const id = createIdForNewReporting(reportings)
 
-      const updatedReportings = [
-        ...reportings,
-        {
-          context: ReportingContext.SIDE_WINDOW,
-          isFormDirty: false,
-          reporting: getReportingInitialValues({ ...response.data, createdAt: new Date().toISOString(), id })
-        }
-      ]
+      const duplicatedReporting = {
+        context: ReportingContext.SIDE_WINDOW,
+        isFormDirty: false,
+        reporting: getReportingInitialValues({ ...response.data, createdAt: new Date().toISOString(), id })
+      }
 
-      await dispatch(
-        multiReportingsActions.setSelectedReportings({ activeReportingId: id, selectedReportings: updatedReportings })
-      )
+      await dispatch(multiReportingsActions.setReporting(duplicatedReporting))
+      await dispatch(multiReportingsActions.setActiveReportingId(id))
 
       await dispatch(
         setReportingFormVisibility({
@@ -36,16 +31,9 @@ export const duplicateReporting = (reportingId: number) => async (dispatch, getS
         })
       )
     } else {
-      throw Error('Erreur à la récupération du signalement')
+      throw Error('Erreur à la duplication du signalement')
     }
   } catch (error) {
     dispatch(setToast({ containerId: 'sideWindow', message: error }))
   }
-
-  dispatch(
-    setReportingFormVisibility({
-      context: ReportingContext.SIDE_WINDOW,
-      visibility: VisibilityState.VISIBLE
-    })
-  )
 }
