+++ frontend/src/domain/use_cases/reportings/switchReporting.ts
@@ -1,33 +1,13 @@
-import { saveReportingInMultiReportingsState } from './saveReportingInMultiReportingsState'
-import { setReportingFormVisibility } from '../../shared_slices/Global'
+import { setReportingFormVisibility, ReportingContext, VisibilityState } from '../../shared_slices/Global'
 import { multiReportingsActions } from '../../shared_slices/MultiReportings'
-import { ReportingContext, VisibilityState } from '../../shared_slices/ReportingState'
 
-export const switchReporting =
-  (nextReportingId: number, reportingContext: ReportingContext) => async (dispatch, getState) => {
-    const {
-      multiReportings: { selectedReportings },
-      reportingState: { reportingState }
-    } = getState()
+export const switchReporting = (nextReportingId: number, reportingContext: ReportingContext) => async dispatch => {
+  dispatch(multiReportingsActions.setActiveReportingId(nextReportingId))
 
-    let updatedReportings = [...selectedReportings]
-
-    // We want to save the active form before switching on another
-    if (reportingState) {
-      updatedReportings = await dispatch(saveReportingInMultiReportingsState())
-    }
-
-    await dispatch(
-      multiReportingsActions.setSelectedReportings({
-        activeReportingId: nextReportingId,
-        selectedReportings: updatedReportings
-      })
-    )
-
-    dispatch(
-      setReportingFormVisibility({
-        context: reportingContext,
-        visibility: VisibilityState.VISIBLE
-      })
-    )
-  }
+  dispatch(
+    setReportingFormVisibility({
+      context: reportingContext,
+      visibility: VisibilityState.VISIBLE
+    })
+  )
+}
