+++ frontend/src/features/Reportings/ReportingsButton/SearchReportings.tsx
@@ -3,8 +3,7 @@ import { useDispatch } from 'react-redux'
 import styled from 'styled-components'
 
 import { sideWindowPaths } from '../../../domain/entities/sideWindow'
-import { setDisplayedItems } from '../../../domain/shared_slices/Global'
-import { ReportingContext } from '../../../domain/shared_slices/ReportingState'
+import { setDisplayedItems, ReportingContext } from '../../../domain/shared_slices/Global'
 import { addReporting } from '../../../domain/use_cases/reportings/addReporting'
 import { useAppSelector } from '../../../hooks/useAppSelector'
 import { sideWindowActions } from '../../SideWindow/slice'
