+++ frontend/src/features/Semaphores/SearchSemaphoreButton/index.tsx
@@ -2,8 +2,7 @@ import { Icon, Size } from '@mtes-mct/monitor-ui'
 import styled from 'styled-components'
 
 import { SearchSemaphores } from './SearchSemaphores'
-import { setDisplayedItems } from '../../../domain/shared_slices/Global'
-import { ReportingContext, VisibilityState } from '../../../domain/shared_slices/ReportingState'
+import { setDisplayedItems, ReportingContext, VisibilityState } from '../../../domain/shared_slices/Global'
 import { reduceReportingFormOnMap } from '../../../domain/use_cases/reportings/reduceReportingFormOnMap'
 import { useAppDispatch } from '../../../hooks/useAppDispatch'
 import { useAppSelector } from '../../../hooks/useAppSelector'
