+++ frontend/src/features/map/overlays/reportings/ReportingCard.tsx
@@ -14,7 +14,8 @@ import { useDispatch } from 'react-redux'
 import styled from 'styled-components'
 
 import { ReportingTypeEnum, getFormattedReportingId, ReportingTypeLabels } from '../../../../domain/entities/reporting'
-import { ReportingContext, reportingStateActions } from '../../../../domain/shared_slices/ReportingState'
+import { ReportingContext } from '../../../../domain/shared_slices/Global'
+import { multiReportingsActions } from '../../../../domain/shared_slices/MultiReportings'
 import { editReportingInLocalStore } from '../../../../domain/use_cases/reportings/editReportingInLocalStore'
 import { useAppSelector } from '../../../../hooks/useAppSelector'
 
@@ -70,7 +71,7 @@ export function ReportingCard({
   }
 
   const closeReportingCard = useCallback(() => {
-    dispatch(reportingStateActions.setSelectedReportingIdOnMap(undefined))
+    dispatch(multiReportingsActions.setSelectedReportingIdOnMap(undefined))
   }, [dispatch])
 
   useEffect(() => {
