+++ frontend/src/features/map/overlays/semaphores/SemaphoreCard.tsx
@@ -4,8 +4,7 @@ import { Tooltip, Whisper } from 'rsuite'
 import styled from 'styled-components'
 
 import { ReportingSourceEnum } from '../../../../domain/entities/reporting'
-import { setOverlayCoordinates } from '../../../../domain/shared_slices/Global'
-import { ReportingContext } from '../../../../domain/shared_slices/ReportingState'
+import { setOverlayCoordinates, ReportingContext } from '../../../../domain/shared_slices/Global'
 import { resetSelectedSemaphore } from '../../../../domain/shared_slices/SemaphoresSlice'
 import { addReporting } from '../../../../domain/use_cases/reportings/addReporting'
 import { useAppDispatch } from '../../../../hooks/useAppDispatch'
