+++ frontend/src/features/map/tools/interest_points/InterestPointMapButton.tsx
@@ -4,13 +4,12 @@ import styled from 'styled-components'
 
 import { EditInterestPoint } from './EditInterestPoint'
 import { MapToolType } from '../../../../domain/entities/map/constants'
-import { setisMapToolVisible } from '../../../../domain/shared_slices/Global'
+import { setisMapToolVisible, ReportingContext, VisibilityState } from '../../../../domain/shared_slices/Global'
 import {
   deleteInterestPointBeingDrawed,
   drawInterestPoint,
   endInterestPointDraw
 } from '../../../../domain/shared_slices/InterestPoint'
-import { ReportingContext, VisibilityState } from '../../../../domain/shared_slices/ReportingState'
 import { reduceReportingFormOnMap } from '../../../../domain/use_cases/reportings/reduceReportingFormOnMap'
 import { useAppDispatch } from '../../../../hooks/useAppDispatch'
 import { useAppSelector } from '../../../../hooks/useAppSelector'
