+++ frontend/src/features/map/tools/measurements/MeasurementMapButton.tsx
@@ -4,9 +4,8 @@ import styled from 'styled-components'
 
 import { CustomCircleRange } from './CustomCircleRange'
 import { MapToolType, MeasurementType } from '../../../../domain/entities/map/constants'
-import { setisMapToolVisible } from '../../../../domain/shared_slices/Global'
+import { setisMapToolVisible, ReportingContext, VisibilityState } from '../../../../domain/shared_slices/Global'
 import { setMeasurementTypeToAdd } from '../../../../domain/shared_slices/Measurement'
-import { ReportingContext, VisibilityState } from '../../../../domain/shared_slices/ReportingState'
 import { reduceReportingFormOnMap } from '../../../../domain/use_cases/reportings/reduceReportingFormOnMap'
 import { useAppDispatch } from '../../../../hooks/useAppDispatch'
 import { useAppSelector } from '../../../../hooks/useAppSelector'
