+++ frontend/src/features/missions/MissionsButton/index.tsx
@@ -4,8 +4,7 @@ import { useDispatch } from 'react-redux'
 import styled from 'styled-components'
 
 import { sideWindowPaths } from '../../../domain/entities/sideWindow'
-import { setDisplayedItems } from '../../../domain/shared_slices/Global'
-import { ReportingContext, VisibilityState } from '../../../domain/shared_slices/ReportingState'
+import { setDisplayedItems, ReportingContext, VisibilityState } from '../../../domain/shared_slices/Global'
 import { addMission } from '../../../domain/use_cases/missions/addMission'
 import { reduceReportingFormOnMap } from '../../../domain/use_cases/reportings/reduceReportingFormOnMap'
 import { useAppSelector } from '../../../hooks/useAppSelector'
