+++ frontend/src/types.ts
@@ -47,3 +47,5 @@ export type MonitorEnum = {
     libelle: string
   }
 }
+
+export type AtLeast<T, K extends keyof T> = Partial<T> & Pick<T, K>
